# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Cards Management API
  description: API для управления банковскими картами
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        201:
          description: Пользователь зарегистрирован
        400:
          description: Невалидные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        200:
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UserRegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        name:
          type: string
          minLength: 2
    UserLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        expiresIn:
          type: integer

    # Cards
...
    ApiError:
      type: object
      properties:
        status:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "Card not found with id: 1"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-06T12:34:56.789Z"